add : ajout de la confirmation de suppression

+++ script.js
@@ -15,6 +15,9 @@ const all_logs_container = document.getElementById("all_transactions_log")
 let transactions = JSON.parse(localStorage.getItem('transactions')) || [];
 let editIndex = null;
 update_balances()
+display_log()
+
+// fonction qui met a jour les 3 dernieres transactions a afficher
 function display_log() {
     log_container.innerHTML = "";
     transactions.slice(-3).reverse().forEach((transaction, i) => {
@@ -28,12 +31,12 @@ function display_log() {
         div.innerHTML = `
     <div class="grid grid-cols-4 w-full">
         <div class="col-span-1 flex flex-col justify-center">
-            <p class="text-black">Le ${transaction.date}</p>
-            <p class="font-bold">${sign}${transaction.amount}DH</p>
+            <p class="text-sm sm:text-base md:text-xl text-black>${transaction.date}</p>
+            <p class="font-bold text-sm sm:text-base md:text-xl">${sign}${transaction.amount}DH</p>
         </div>
         <div class="col-span-2 flex flex-col justify-center">
-            <p class="text-black">Description :</p>
-            <p id="log_description">${transaction.description}</p>
+            <p class="text-sm sm:text-base md:text-xl text-black">Description :</p>
+            <p>${transaction.description}</p>
         </div>
         <div class="flex gap-4 justify-center items-center">
             <button class="edit_button" data-index="${idx}"><img class="w-10" src="images/edit.svg" alt="Edit"></button>
@@ -41,27 +44,65 @@ function display_log() {
         </div>
     </div>
 `
-            ;
-        log_container.appendChild(div);
+        log_container.appendChild(div)
     });
 }
-display_log()
+//fonction qui met a jour et affiche toutes les transactions
+function display_all_logs() {
+    all_logs_container.innerHTML = ""
+    transactions.slice().reverse().forEach((transaction, i) => {
+        // Index réel dans le tableau principal car affiché en reverse
+        let idx = transactions.length - 1 - i;
+        let sign = transaction.type === "income" ? "+" : "-";
+        const div = document.createElement("div")
+        let color = transaction.type === "income" ? "#02D302" : "#FF4E4E"
+        div.style.backgroundColor = color
+        div.className = "border-2 border-white rounded-lg w-3/4 h-16 text-xl"
+        div.innerHTML = `
+    <div class="grid grid-cols-4 w-full">
+        <div class="col-span-1 flex flex-col justify-center">
+            <p class="text-sm sm:text-base md:text-xl text-black">${transaction.date}</p>
+            <p class="font-bold text-sm sm:text-base md:text-xl">${sign}${transaction.amount}DH</p>
+        </div>
+        <div class="col-span-2 flex flex-col justify-center ml-4">
+            <p class="text-sm sm:text-base md:text-xl text-black">Description :</p>
+            <p class="text-sm sm:text-base md:text-xl">${transaction.description}</p>
+        </div>
+        <div class="flex gap-4 justify-center items-center">
+            <button class="edit_button" data-index="${idx}"><img class="w-10" src="images/edit.svg" alt="Edit"></button>
+            <button class="delete_button" data-index="${idx}"><img class="w-10" src="images/trash.svg" alt="Trash"></button>
+        </div>
+    </div>
+`
+
+        all_logs_container.appendChild(div)
+    })
+}
+// fonction qui calcule les soldes positifs , negatif et la balance finale, et les met a jour dans le local storage
 function update_balances() {
     let total_income = transactions
         .filter(t => t.type === "income")
-        .reduce((sum, t) => sum + Number(t.amount), 0);
+        .reduce((sum, t) => sum + Number(t.amount), 0)
     let total_expense = transactions
         .filter(t => t.type === "expense")
-        .reduce((sum, t) => sum + Number(t.amount), 0);
-    let balance = total_income - total_expense;
+        .reduce((sum, t) => sum + Number(t.amount), 0)
+    let balance = total_income - total_expense
 
-    document.getElementById("solde_positif").innerText = "+" + total_income + "DH";
-    document.getElementById("solde_negatif").innerText = "-" + total_expense + "DH";
-    document.getElementById("balance").innerText = balance + "DH";
-    localStorage.setItem("total_income", total_income);
-    localStorage.setItem("total_expense", total_expense);
-    localStorage.setItem("balance", balance);
+    document.getElementById("solde_positif").innerText = "+" + total_income + "DH"
+    document.getElementById("solde_negatif").innerText = "-" + total_expense + "DH"
+    document.getElementById("balance").innerText = balance + "DH"
+    localStorage.setItem("total_income", total_income)
+    localStorage.setItem("total_expense", total_expense)
+    localStorage.setItem("balance", balance)
 }
+
+// listener sur tous les boutons edit et delete
+document.addEventListener("click", function (event) {
+    if (event.target.closest(".edit_button") || event.target.closest(".delete_button")) {
+        edit_delete_buttons(event);
+    }
+})
+// fonction qui modifie ou supprime le bouton écouté
 function edit_delete_buttons() {
     const edit_btn = event.target.closest(".edit_button")
     const delete_btn = event.target.closest(".delete_button")
@@ -75,14 +116,18 @@ function edit_delete_buttons() {
         return
     }
     if (delete_btn) {
-        let idx = Number(delete_btn.dataset.index)
-        transactions.splice(idx, 1)
-        localStorage.setItem("transactions", JSON.stringify(transactions))
-        update_balances()
-        display_log()
-        return
+        if (confirm("Do you really want to delete this transaction ?")) {
+            let idx = Number(delete_btn.dataset.index)
+            transactions.splice(idx, 1)
+            localStorage.setItem("transactions", JSON.stringify(transactions))
+            update_balances()
+            display_all_logs()
+            display_log()
+            return
+        }
     }
 }
+// quand l'utilisateur submit les modifications de sa transaction, récupération des infos dans le local storage
 popup_edit.addEventListener("submit", (event) => {
     event.preventDefault()
     let transaction = {
@@ -99,7 +144,12 @@ popup_edit.addEventListener("submit", (event) => {
     popup_edit.classList.add("hidden")
     editIndex = null
 })
+// bouton close de la popup edit
+close_btn_edit.addEventListener("click", (event) => {
+    popup_edit.classList.add("hidden")
+})
 
+// quand l'utilisateur submit son income , récupérations des infos dans le local storage
 popup_add_income.addEventListener("submit", (event) => {
     event.preventDefault()
     let transaction = {
@@ -113,7 +163,17 @@ popup_add_income.addEventListener("submit", (event) => {
     update_balances()
     display_log()
 })
+//si l'utilisateur clique sur add an income, affichage de la pop up pour ajouter un revenu
+add_income_btn.addEventListener("click", () => {
+    popup_add_income.classList.remove("hidden")
+})
+// bouton close de la popup add an income
+close_btn.addEventListener("click", (event) => {
+    popup_add_income.classList.add("hidden")
+    event.preventDefault()
+})
 
+//quand l'utilisateur submit son expense, récupérations des infos dans le local storage
 popup_add_expense.addEventListener("submit", (event) => {
     event.preventDefault()
     let transaction = {
@@ -123,70 +183,29 @@ popup_add_expense.addEventListener("submit", (event) => {
         type: "expense"
     }
     transactions.push(transaction)
+    localStorage.setItem("transactions", JSON.stringify(transactions));
     update_balances()
     display_log()
 
 })
-
-close_btn.addEventListener("click", (event) => {
-    popup_add_income.classList.add("hidden")
-    event.preventDefault()
-})
-close_btn_edit.addEventListener("click", (event) => {
-    popup_edit.classList.add("hidden")
-})
-show_all_close_btn.addEventListener("click", (event) => {
-    popup_show.classList.add("hidden")
-    event.preventDefault()
-})
-
-show_all_btn.addEventListener("click", (event) => {
-    popup_show.classList.remove("hidden")
-    event.preventDefault()
-    all_logs_container.innerHTML = ""
-    transactions.slice().reverse().forEach((transaction, i) => {
-        // Calcul de l'index réel dans le tableau principal
-        let idx = transactions.length - 1 - i;
-        let sign = transaction.type === "income" ? "+" : "-";
-        const div = document.createElement("div")
-        let color = transaction.type === "income" ? "#02D302" : "#FF4E4E"
-        div.style.backgroundColor = color
-        div.className = "border-2 border-white rounded-lg w-3/4 h-16 text-xl"
-        div.innerHTML = `
-    <div class="grid grid-cols-4 w-full">
-        <div class="col-span-1 flex flex-col justify-center">
-            <p class="text-black">Le ${transaction.date}</p>
-            <p class="font-bold">${sign}${transaction.amount}DH</p>
-        </div>
-        <div class="col-span-2 flex flex-col justify-center">
-            <p class="text-black">Description :</p>
-            <p id="log_description">${transaction.description}</p>
-        </div>
-        <div class="flex gap-4 justify-center items-center">
-            <button class="edit_button" data-index="${idx}"><img class="w-10" src="images/edit.svg" alt="Edit"></button>
-            <button class="delete_button" data-index="${idx}"><img class="w-10" src="images/trash.svg" alt="Trash"></button>
-        </div>
-    </div>
-`
-
-        all_logs_container.appendChild(div);
-    })
-
-})
-add_income_btn.addEventListener("click", () => {
-    popup_add_income.classList.remove("hidden")
-})
+// si l'utilisateur clique sur add an expense, affichage de la popup pour ajouter une dépense
 expense_btn.addEventListener("click", () => {
     popup_add_expense.classList.remove("hidden")
 })
+//bouton close de la popup add an expense
 close_btn_expense.addEventListener("click", (event) => {
     popup_add_expense.classList.add("hidden");
     event.preventDefault();
 })
 
-document.addEventListener("click", function (event) {
-    if (event.target.closest(".edit_button") || event.target.closest(".delete_button")) {
-        edit_delete_buttons(event);
-    }
+// si l'utilisateur clique sur show all transactions
+show_all_btn.addEventListener("click", (event) => {
+    popup_show.classList.remove("hidden")
+    event.preventDefault()
+    display_all_logs()
 })
-
+// bouton close de la popup show all transactions
+show_all_close_btn.addEventListener("click", (event) => {
+    popup_show.classList.add("hidden")
+    event.preventDefault()
+})
